Some python cleanup
