+++ panda_agent/agent.py
@@ -1,15 +1,5 @@
-import grpc
-import concurrent.futures as futures
-import queue
-
 from pandare import Panda
-
-import panda_interface_pb2_grpc as pb_grpc
-import panda_interface_pb2 as pb
-
-PORT = '[::]:50051'
-
-executor = futures.ThreadPoolExecutor(max_workers=10)
+import queue
 
 class PandaAgent:
     # sentinel object to signal end of queue
@@ -73,43 +63,3 @@ class PandaAgent:
         # Receive message from queue
         resp = retQueue.get(block=True, timeout=None)
         return resp
-
-class PandaExecutorServicer(pb_grpc.PandaExecutorServicer):
-    def __init__(self, server, agent: PandaAgent):
-        self.server = server
-        self.agent = agent
-        self.agent_thread = None
-    
-    def isAgentRunning(self):
-        return self.agent_thread is not None
-    
-    def BootMachine(self, request, context):
-        if self.agent.hasStarted:
-            return
-
-        # start panda in a new thread, because qemu blocks this thread otherwise
-        executor.submit(self.agent.start)
-        return pb.BootMachineReply()
-    
-    def Shutdown(self, request, context):
-        self.agent.stop()
-        self.server.stop(grace=5)
-        return pb.ShutdownReply()
-    
-    def RunCommand(self, request, context):
-        output = self.agent.run_command(request.command)
-        return pb.RunCommandReply(statusCode=0, output=output)
-
-def serve():
-    panda = Panda(generic='x86_64')
-    agent = PandaAgent(panda)
-    server = grpc.server(executor)
-    pb_grpc.add_PandaExecutorServicer_to_server(
-        PandaExecutorServicer(server, agent), server)
-    server.add_insecure_port(PORT)
-    print(f'panda agent grpc server listening on port {PORT}')
-    server.start()
-    server.wait_for_termination()
-
-if __name__ == "__main__":
-    serve()
