+++ panda_agent/generate_protobuf.sh
@@ -1,20 +1,20 @@
 #!/bin/bash
 
-# Generate golang stubs
-# protoc --go-out=. --go_opt=paths=source_relative proto/panda_interface.proto
-python3 -m grpc_tools.protoc --python_out=proto --pyi_out=proto --grpc_python_out=proto -I./proto ./proto/panda_interface.proto
-
+rm -rf pb/ *_pb2*
 mkdir -p pb
+
+PYTHON_OUT=pb
+GO_OUT=pb
 
 # Generate python stubs
 # pip3 install grpcio-tools
 python3 -m grpc_tools.protoc \
-    --python_out=. \
-    --pyi_out=. \
-    --grpc_python_out=. \
-    --go_out=pb \
+    --python_out=$PYTHON_OUT \
+    --pyi_out=$PYTHON_OUT \
+    --grpc_python_out=$PYTHON_OUT \
+    --go_out=$GO_OUT \
     --go_opt=paths=source_relative \
-    --go-grpc_out=pb \
+    --go-grpc_out=$GO_OUT \
     --go-grpc_opt=paths=source_relative \
     -I./proto \
-    ./proto/panda_interface.proto \
+    ./proto/*.proto \
